move editor theme selector into settings tab
